pyproject.toml
[tool.poetry]
name = "extract-data-domsistemas-py"
version = "0.1.0"
description = ""
authors = ["Elder Vivot <eldervivot@gmail.com>"]
readme = "README.md"
packages = [{include = "extract_data_domsistemas_py"}]

[tool.poetry.dependencies]
python = "^3.9"
pyodbc = "^4.0.35"
pandas = "^1.5.3"
python-dotenv = "^1.0.0"
flask = "^2.2.3"
flask-apscheduler = "^1.12.4"
aiohttp = "^3.8.4"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
mypy = "^1.1.1"
flake8 = "^6.0.0"
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
